Refactored + commented fraction strategy pattern
